Refactor review rating model to use int for rating and update profile screen to fetch reviews correctly

+++ lib/review_rating/models/review_rating.dart
@@ -26,7 +26,7 @@ class ReviewRating {
 }
 
 class Fields {
-    double rating;
+    int rating;
     String review;
     String id;
     bool canDelete;
